+++ CHANGELOG.md
@@ -1,5 +1,41 @@
 # cypheir-color-scheme
 
+## 0.6.5
+
+### Patch Changes
+
+- CMYK Converters
+
+## 0.6.4
+
+### Patch Changes
+
+- Colour Lookup interface
+
+## 0.6.3
+
+### Patch Changes
+
+- Colour Lookup method tweak
+
+## 0.6.2
+
+### Patch Changes
+
+- package.json dependencies
+
+## 0.6.1
+
+### Patch Changes
+
+- Bundle Dependencies
+
+## 0.6.0
+
+### Minor Changes
+
+- Colour Name Finder
+
 ## 0.5.3
 
 ### Patch Changes